Various bug fixes
I made sure that the functions were returning the proper types.
I also fixed a bug in the humidity init code that sets up the factory
programmed values.

+++ TempHumiditySensor/inc/i2c_interface.h
@@ -9,6 +9,6 @@
 #define INC_I2C_INTERFACE_H_
 
 void initialize_i2c();
-uint32_t i2c_read_byte(uint8_t slave_address, uint8_t register_address);
+uint8_t i2c_read_byte(uint8_t slave_address, uint8_t register_address);
 
 #endif /* INC_I2C_INTERFACE_H_ */
